+++ changelog.md
@@ -2,7 +2,8 @@
 ## v0.11
 * Multi-devices: a single device object acts as abstraction to multiple devices
 * DDwrt Wi-Fi based presence detection
-* Fix broken Chromecast controls
+* Fix broken Chromecast controls (partially)
+* Allow driver modules to define actions in a class attribute
 
 ## v0.10
 * Chainable actions